添加demo

+++ csvReader.h
@@ -18,7 +18,7 @@ typedef enum type //sortation means the priority
 typedef union _element //cell的数据用此类型
 {
     char char_ch[200];//不能超过99个字节
-    size_t integer_num;
+    long long integer_num;
     double float_num;
     char nan[4];
 }element;
